queue_services/account-mailer/pyproject.toml
[tool.poetry]
name = "account-mailer"
version = "0.2.0"
description = ""
authors = ["Avni Salhotra <avni.salhotra@gov.bc.ca>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
blinker = "1.8.2"
charset-normalizer = "3.3.2"
click = "8.1.7"
expiringdict = "1.2.2"
google-api-core = "1.34.1"
google-auth = "2.28.2"
google-cloud-core = "2.4.3"
google-cloud-pubsub = "2.20.2"
google-cloud-storage = "2.14.0"
googleapis-common-protos = "1.63.0"
grpc-google-iam-v1 = "0.13.0"
grpcio-status = "1.48.2"
grpcio = "1.64.0"
idna = "3.7"
importlib-resources = "5.13.0"
jaeger-client = "4.8.0"
msgpack = "1.0.8"
opentracing = "2.4.0"
pkgutil-resolve-name = "1.3.10"
protobuf = ">=3.19.5,<3.20.0"
pyrfc3339 = "1.1"
pycountry = "23.12.11"
pyrsistent = "0.20.0"
pytz = "^2024.1"
semver = "3.0.2"
six = "1.16.0"
threadloop = "1.0.2"
thrift = "0.20.0"
tornado = "6.4.2"
urllib3 = "2.2.2"
zipp = "3.19.1"

# VCS dependencies
auth-api = { git = "https://github.com/bcgov/sbc-auth.git", rev = "dependency_upgrades_p2", subdirectory = "auth-api" }
simple-cloudevent = { git = "https://github.com/daxiom/simple-cloudevent.py.git" }
cloud-sql-python-connector = "^1.13.0"

[tool.poetry.group.dev.dependencies]
psycopg2 = "2.9.9"
pytest = "^8.3.2"
pytest-mock = "^3.14.0"
pyhamcrest = "^2.1.0"
pytest-cov = "^5.0.0"
freezegun = "^1.5.1"
flake8 = "5.0.4"
flake8-blind-except = "^0.2.1"
flake8-debugger = "^4.1.2"
flake8-docstrings = "^1.7.0"
flake8-isort = "^6.1.1"
flake8-print = "^5.0.0"
flake8-quotes = "^3.4.0"
pep8-naming = "^0.14.1"
coverage = "^7.6.1"
pylint = "3.2.3"
pylint-flask = "^0.6"
pydocstyle = "^6.3.0"
isort = "^5.13.2"
lovely-pytest-docker = "^1.0.0"
pytest-asyncio = "0.18.3"
astroid = "^3.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
